Moderation reports: Create ReportUtil class, add category name function
