+++ src/Module/Moderation/Reports.php
@@ -19,7 +19,7 @@ use Friendica\Core\Session\Capability\IHandleUserSessions;
 use Friendica\Database\Database;
 use Friendica\Database\DBA;
 use Friendica\Module\BaseModeration;
-use Friendica\Moderation\Entity\Report;
+use Friendica\Module\Moderation\Utils\ReportUtil;
 use Friendica\Module\Response;
 use Friendica\Navigation\SystemMessages;
 use Friendica\Util\DateTimeFormat;
@@ -30,39 +30,21 @@ class Reports extends BaseModeration
 {
 	/** @var Database */
 	private $database;
+	/** @var ReportUtil */
+	protected $reportUtil;
 
-	public function __construct(Database $database, Page $page, AppHelper $appHelper, SystemMessages $systemMessages, IHandleUserSessions $session, L10n $l10n, BaseURL $baseUrl, Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
+	public function __construct(Database $database, Page $page, ReportUtil $reportUtil, AppHelper $appHelper, SystemMessages $systemMessages, IHandleUserSessions $session, L10n $l10n, BaseURL $baseUrl, Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
 	{
 		parent::__construct($page, $appHelper, $systemMessages, $session, $l10n, $baseUrl, $args, $logger, $profiler, $response, $server, $parameters);
 
-		$this->database = $database;
+		$this->database   = $database;
+		$this->reportUtil = $reportUtil;
 	}
 
 	protected function post(array $request = [])
 	{
 		// @todo check if POST is really used here
 		$this->content($request);
-	}
-
-	// A copy of the one in Report/Create.php - please consolidate!
-	public function getReportCategoryName(int $category): string
-	{
-		switch ($category) {
-			case Report::CATEGORY_SPAM:
-				return $this->t('Spam');
-			case Report::CATEGORY_ILLEGAL:
-				return $this->t('Illegal Content');
-			case Report::CATEGORY_SAFETY:
-				return $this->t('Community Safety');
-			case Report::CATEGORY_UNWANTED:
-				return $this->t('Unwanted Content/Behavior');
-			case Report::CATEGORY_VIOLATION:
-				return $this->t('Rules Violation');
-			case Report::CATEGORY_OTHER:
-				return $this->t('Other');
-			default:
-				return "";
-		};
 	}
 
 	protected function content(array $request = []): string
@@ -96,7 +78,7 @@ LIMIT ?, ?",
 		while ($report = $this->database->fetch($query)) {
 			$report['posts']    = [];
 			$report['created']  = DateTimeFormat::local($report['created'], DateTimeFormat::MYSQL);
-			$report['category'] = self::getReportCategoryName($report['category-id']);
+			$report['category'] = $this->reportUtil->getReportCategoryName($report['category-id']);
 
 			$reports[$report['id']] = $report;
 		}
